+++ packages/antd/modules/widgets/core/Button.jsx
@@ -1,53 +1,47 @@
 import React from "react";
 import { Button } from "antd";
-import { PlusOutlined, PlusCircleOutlined, DeleteFilled } from "@ant-design/icons";
 
-export default ({type, onClick, label, readonly, config: {settings}}) => {
-  const hideLabelsFor = {
-    "addRuleGroup": true
-  };
-  const btnLabel = hideLabelsFor[type] ? "" : label;
-  const hasLabel = !!btnLabel;
+const hideLabelsFor = {
+  "addRuleGroup": true,
+  "delGroup": true,
+  "delRuleGroup": true,
+  "delRule": true,
+};
 
-  const typeToIcon = {
-    "addRule": <PlusOutlined />,
-    "addGroup": <PlusCircleOutlined />,
-    "delRule": <DeleteFilled />, //?
-    "delGroup": <DeleteFilled />,
-    "delRuleGroup": <DeleteFilled />,
-    "addRuleGroup": <PlusOutlined />,
-  };
+const typeToClass = {
+  "addRule": "action action--ADD-RULE",
+  "addGroup": "action action--ADD-GROUP",
+  "delRule": "action action--DELETE",
+  "delGroup": "action action--DELETE",
+  "delRuleGroup": "action action--DELETE",
+  "addRuleGroup": "action action--ADD-RULE",
+};
 
-  const typeToClass = {
-    "addRule": "action action--ADD-RULE",
-    "addGroup": "action action--ADD-GROUP",
-    "delRule": "action action--DELETE", //?
-    "delGroup": "action action--DELETE",
-    "delRuleGroup": "action action--DELETE",
-    "addRuleGroup": <PlusOutlined />,
-  };
+const typeToType = {
+  "delRule": "text",
+  // "delGroup": "default",
+  // "delRuleGroup": "default",
+};
 
-  const typeToType = {
-    "delRule": "text",
-    // "delGroup": "default",
-    // "delRuleGroup": "default",
-  };
+const dangerFor = {
+  "delRule": true,
+  "delGroup": true,
+  "delRuleGroup": true,
+};
 
-  const dangerFor = {
-    "delRule": true,
-    "delGroup": true,
-    "delRuleGroup": true,
-  };
-
+export default (props) => {
+  const {type, onClick, label, readonly, config: {settings}, renderIcon} = props;
   const {renderSize} = settings;
 
+  const icon = renderIcon?.(props);
+  const btnLabel = hideLabelsFor[type] ? "" : label;
 
   return (
     <Button
       danger={dangerFor[type] === true}
       key={type}
       type={typeToType[type] || "default"}
-      icon={typeToIcon[type]}
+      icon={icon}
       className={typeToClass[type]}
       onClick={onClick}
       size={renderSize}
