+++ packages/antd/modules/widgets/core/Icon.jsx
@@ -1,0 +1,24 @@
+import React from "react";
+import { PlusOutlined, PlusCircleOutlined, DeleteFilled, HolderOutlined } from "@ant-design/icons";
+import { Utils } from "@react-awesome-query-builder/ui";
+const { DragIcon } = Utils;
+
+const typeToIcon = {
+  "addRule": <PlusOutlined />,
+  "addGroup": <PlusCircleOutlined />,
+  "delRule": <DeleteFilled />,
+  "delGroup": <DeleteFilled />,
+  "delRuleGroup": <DeleteFilled />,
+  "addRuleGroup": <PlusOutlined />,
+  "addRuleGroupExt": <PlusOutlined />,
+  "drag": <HolderOutlined />,
+};
+
+export default ({type}) => {
+  let icon = typeToIcon[type];
+  if (!icon && type === "drag") {
+    icon = <DragIcon />;
+  }
+
+  return icon || null;
+};
